wip adding jest tests -- need to break up to test
